refactor: modelmapper사용, 주소 long타입 변경, 불필요 주석 삭제, api에 user추가

+++ src/main/java/egenius/user/entity/User.java
@@ -31,14 +31,7 @@ public class User extends BaseTimeEntity implements UserDetails {
     @Column(name = "phone_number", length = 15)
     private String phoneNumber;
     @Column(name = "address", columnDefinition = "int default 1")
-    private int address;
-
-    /**
-     * 의존성 역전
-     * Entity -> Domain
-     *
-     * @return
-     */
+    private Long address;
 
     public void hashPassword(String password) {
         //      this.password = password;
